server/python/requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
transformers==4.30.2
spacy==3.6.1
torch==2.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
